% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crs_cpp}
\alias{crs_cpp}
\title{C++ reimplementation of survPen::crs, including derivatives.}
\usage{
crs_cpp(
  x,
  df = 10L,
  knots = NULL,
  derivs = 0L,
  intercept = FALSE,
  survPen_compatible = FALSE
)
}
\arguments{
\item{x}{a vec of values to provide}

\item{df}{a non-negative integer for the degrees of freedom (ignored if
the knots are specified)}

\item{knots}{a vec of knots (defaults to NULL)}

\item{derivs}{a non-negative integer for the derivative (default 0)}

\item{intercept}{a bool for whether to include the intercept}

\item{survPen_compatible}{a bool for whether to use \code{unique(x)} for determining the knots if the \code{knots} are not specified}
}
\value{
an \code{Rcpp::NumericMatrix} with attributes \code{class="crs_cpp"},
\code{pen} for the penalty matrix, \code{knots} for the knots, \code{derivs} for the order of
derivatives, and \code{intercept} for a bool/logical for whether the intercept is
included.
}
\description{
The re-implementation uses the same algorithm but uses
different defaults, and allows for derivatives.
For the changes in defaults: the function here assumes that
the intercept is false (as per \code{splines::ns}); when the knots are
not specified, the function defaults to using \code{x} rather than \code{unique(x)}
for the quantiles; and the \code{df} argument comes before the \code{knots} argument.
To use the \code{survPen::crs} defaults in R, use
\code{crs_cpp(..., intercept=TRUE, survPen_compatible=TRUE)}.
}
\examples{
x     <- seq(0, 1, length.out = 11)
knots <- c(1, 3, 5, 7) / 10
crs1  <- crs_cpp(x, knots = knots)
dim(crs1)
# Comparison with survPen::crs requires the survPen package:
\dontrun{
library(survPen)
crs2 <- crs(x, knots, intercept = FALSE)
range(crs1 - crs2$bs)
range(attr(crs1, "pen") - crs2$pen)
}
}
