% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.summary.genhaz_fit}
\alias{print.summary.genhaz_fit}
\title{Print a summary of a fitted GH model}
\usage{
\method{print}{summary.genhaz_fit}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{"summary.genhaz_fit"} object from \code{summary.genhaz_fit()}.}

\item{digits}{Number of significant digits. Default 4.}

\item{...}{Currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a summary of a fitted GH model
}
\examples{
\dontrun{
print(summary(fit))
}
}
