% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantile_type7}
\alias{quantile_type7}
\title{Armadillo-based reimplementation of \code{stats::quantile()} for \code{type=7}.}
\usage{
quantile_type7(x, probs)
}
\arguments{
\item{x}{a vec of values for which to find the quantiles}

\item{probs}{a vec of probilities that specify which quantiles}
}
\value{
a vec of quantiles
}
\description{
Allows for NAs.
}
\examples{
set.seed(12345)
y <- rnorm(1e5)
# Should be near zero for all quantiles
quantile(y) - quantile_type7(y, seq(0, 1, length.out = 5))
quantile(y, 0.975) - quantile_type7(y, 0.975)
# NA probs are allowed (unlike base quantile()):
quantile_type7(y, probs = c(0.1, 0.5, 1, 2, 5, 10, 50, NA) / 100)
}
